added lib for nearfull disk data storage
