+++ tests/iostability/test_iostability_degraded_path.py
@@ -25,14 +25,14 @@ from datetime import timedelta
 import pytest
 
 from commons import configmanager
-from commons.constants import K8S_SCRIPTS_PATH, POD_NAME_PREFIX
+from commons.constants import K8S_SCRIPTS_PATH
 from commons.helpers.health_helper import Health
 from commons.helpers.pods_helper import LogicalNode
 from commons.params import LATEST_LOG_FOLDER
 from commons.utils import assert_utils, support_bundle_utils
 from config import CMN_CFG
 from conftest import LOG_DIR
-from libs.durability.disk_failure_recovery_libs import DiskFailureRecoveryLib
+from libs.prefill.near_full_data_storage import DiskNearFullStorage
 from libs.ha.ha_common_libs_k8s import HAK8s
 from libs.iostability.iostability_lib import IOStabilityLib
 from libs.s3.s3_test_lib import S3TestLib
@@ -63,7 +63,6 @@ class TestIOWorkloadDegradedPath:
             else:
                 cls.worker_node_list.append(node_obj)
         cls.ha_obj = HAK8s()
-        cls.dfr = DiskFailureRecoveryLib()
         cls.test_cfg = configmanager.get_config_wrapper(fpath="config/iostability_test.yaml")
         cls.setup_type = CMN_CFG["setup_type"]
         cls.s3t_obj = S3TestLib()
@@ -140,18 +139,18 @@ class TestIOWorkloadDegradedPath:
         client = len(self.worker_node_list) * self.test_cfg['sessions_per_node_vm']
         percentage = self.test_cfg['nearfull_storage_percentage']
         self.log.info("Step 1: calculating byte count for required percentage")
-        resp = self.dfr.get_user_data_space_in_bytes(master_obj=self.master_node_list[0],
-                                                     memory_percent=percentage)
+        resp = DiskNearFullStorage.get_user_data_space_in_bytes(master_obj=self.master_node_list[0],
+                                                                memory_percent=percentage)
         assert_utils.assert_true(resp[0], resp[1])
         self.log.info("Need to add %s bytes for required percentage", resp[1])
         self.log.info("Step 2: performing writes till we reach required percentage")
 
-        ret = DiskFailureRecoveryLib.perform_near_full_sys_writes(s3userinfo=s3userinfo,
-                                                                  user_data_writes=int(resp[1]),
-                                                                  bucket_prefix=bucket_prefix,
-                                                                  client=client)
+        ret = DiskNearFullStorage.perform_near_full_sys_writes(s3userinfo=s3userinfo,
+                                                               user_data_writes=int(resp[1]),
+                                                               bucket_prefix=bucket_prefix,
+                                                               client=client)
         assert_utils.assert_true(ret[0], ret[1])
-        for each in ret:
+        for each in ret[1]:
             each["num_clients"] = (len(self.worker_node_list) - 1) \
                                   * self.test_cfg['sessions_per_node_vm']
         self.log.debug("write operation data: %s", ret)
@@ -167,12 +166,11 @@ class TestIOWorkloadDegradedPath:
         while datetime.now() < end_time:
             loop += 1
             self.log.info("%s remaining time for reading loop", (end_time - datetime.now()))
-            read_ret = DiskFailureRecoveryLib.perform_near_full_sys_operations(
-                s3userinfo=s3userinfo,
-                workload_info=ret[1],
-                skipread=False,
-                validate=True,
-                skipcleanup=True)
+            read_ret = DiskNearFullStorage.perform_near_full_sys_operations(s3userinfo=s3userinfo,
+                                                                            workload_info=ret[1],
+                                                                            skipread=False,
+                                                                            validate=True,
+                                                                            skipcleanup=True)
             self.log.info("%s interation is done", loop)
             assert_utils.assert_true(read_ret[0], read_ret[1])
         # To Do delete operation, will be enabled after support from cortx
