+++ tests/iostability/test_iostability_happy_path.py
@@ -30,7 +30,7 @@ from commons.params import LATEST_LOG_FOLDER
 from commons.utils import support_bundle_utils, assert_utils
 from config import CMN_CFG
 from conftest import LOG_DIR
-from libs.durability.disk_failure_recovery_libs import DiskFailureRecoveryLib
+from libs.prefill.near_full_data_storage import DiskNearFullStorage
 from libs.iostability.iostability_lib import IOStabilityLib
 from libs.s3 import ACCESS_KEY
 from libs.s3 import SECRET_KEY
@@ -58,7 +58,6 @@ class TestIOWorkload:
                 cls.worker_node_list.append(node_obj)
         cls.test_cfg = configmanager.get_config_wrapper(fpath="config/iostability_test.yaml")
         cls.setup_type = CMN_CFG["setup_type"]
-        cls.dfr = DiskFailureRecoveryLib()
         cls.test_completed = False
         cls.iolib = IOStabilityLib()
 
@@ -107,15 +106,15 @@ class TestIOWorkload:
         client = len(self.worker_node_list) * self.test_cfg['sessions_per_node_vm']
         percentage = self.test_cfg['nearfull_storage_percentage']
         self.log.info("Step 1: calculating byte count for required percentage")
-        resp = self.dfr.get_user_data_space_in_bytes(master_obj=self.master_node_list[0],
-                                                     memory_percent=percentage)
+        resp = DiskNearFullStorage.get_user_data_space_in_bytes(master_obj=self.master_node_list[0],
+                                                                memory_percent=percentage)
         assert_utils.assert_true(resp[0], resp[1])
         self.log.info("Need to add %s bytes for required percentage", resp[1])
         self.log.info("Step 2: performing writes till we reach required percentage")
-        ret = DiskFailureRecoveryLib.perform_near_full_sys_writes(s3userinfo=s3userinfo,
-                                                                  user_data_writes=int(resp[1]),
-                                                                  bucket_prefix=bucket_prefix,
-                                                                  client=client)
+        ret = DiskNearFullStorage.perform_near_full_sys_writes(s3userinfo=s3userinfo,
+                                                               user_data_writes=int(resp[1]),
+                                                               bucket_prefix=bucket_prefix,
+                                                               client=client)
         assert_utils.assert_true(ret[0], ret[1])
         self.log.debug("write operation data: %s", ret)
         self.log.info("Step 3: performing read operations.")
@@ -124,17 +123,18 @@ class TestIOWorkload:
         while datetime.now() < end_time:
             loop += 1
             self.log.info("%s remaining time for reading loop", (end_time - datetime.now()))
-            read_ret = DiskFailureRecoveryLib.perform_near_full_sys_operations(
-                                                                    s3userinfo=s3userinfo,
-                                                                    workload_info=ret[1],
-                                                                    skipread=False,
-                                                                    validate=True,
-                                                                    skipcleanup=True)
+            read_ret = DiskNearFullStorage.perform_near_full_sys_operations(s3userinfo=s3userinfo,
+                                                                            workload_info=ret[1],
+                                                                            skipread=False,
+                                                                            validate=True,
+                                                                            skipcleanup=True)
             self.log.info("%s interation is done", loop)
             assert_utils.assert_true(read_ret[0], read_ret[1])
         self.log.info("Step 4: performing delete operations.")
-        del_ret = DiskFailureRecoveryLib.perform_near_full_sys_operations(
-            s3userinfo=s3userinfo, workload_info=ret[1], skipread=True, validate=False,
-            skipcleanup=False)
+        del_ret = DiskNearFullStorage.perform_near_full_sys_operations(s3userinfo=s3userinfo,
+                                                                       workload_info=ret[1],
+                                                                       skipread=True,
+                                                                       validate=False,
+                                                                       skipcleanup=False)
         assert_utils.assert_true(del_ret[0], del_ret[1])
         self.log.info("ENDED: Perform disk storage near full once and read in loop for 30 days")
